pyproject.toml
[tool.poetry]
name        = "trz-py-utils"
version     = "0.0.0"
description = "Trapezoid Python Utility Library"
authors = ["Justin Mai <justin@trapezoidtech.io>"]
# packages    = [{include = "src"}]

[tool.poetry.dependencies]
python                    = "3.11.6"
aws-requests-auth         = "0.4.3"
aws-lambda-powertools     = "2.26.0"
boto3                     = "1.28.83"
fastjsonschema            = "2.19.0"
# moto                      = "4.2.8"
psycopg2-binary           = "2.9.9"
# poetry-dynamic-versioning = "^1.2.0"
mypy-boto3-lambda         = "^1.34.0"
xdoctest                  = "^1.1.2"
pillow                    = "10.1.0"
mypy-boto3-s3             = "1.34.14"
tqdm                      = "4.66.1"
chardet                   = "5.2.0"

[tool.poetry.group.docs.dependencies]
sphinx           = "5.3.0"
sphinx-rtd-theme = "1.3.0"

[tool.poetry.group.tests.dependencies]
pytest = "7.1.3"
flake8 = "5.0.4"

[tool.poetry.group.dev.dependencies]
ipython           = "8.19.0"

[tool.poetry-dynamic-versioning]
enable = true
vcs = "git"

[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning"]
build-backend = "poetry_dynamic_versioning.backend"
